batch_bot_2/batch_bot/requirements.txt
# ─────────────────────────────────────────────────────────────
#  requirements.txt
#  Install all dependencies:
#    pip install -r requirements.txt
# ─────────────────────────────────────────────────────────────

# Browser automation
selenium>=4.18.0

# Excel / CSV handling
pandas>=2.2.0
openpyxl>=3.1.2

# Captcha (OCR mode) — optional but recommended
pytesseract>=0.3.10
Pillow>=10.2.0

# HTTP (API captcha mode)
requests>=2.31.0

# tkinter ships with Python — no pip install needed
# If missing on Linux: sudo apt install python3-tk
